pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "loona-wakeup"
version = "0.1.0"
description = "Loona no-wake-word wakeup monitor"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6",
    "PyYAML>=6.0",
    "numpy>=1.26",
    "opencv-python>=4.9",
    "sounddevice>=0.4",
    "mediapipe>=0.10.14",
]

[project.scripts]
loona-wakeup = "loona_wakeup.app:run"

[tool.setuptools.packages.find]
where = ["src"]
